\name{cube}
\alias{cube}
cubic value
\title{
cube calculation
}
\description{
cube cal
}
\usage{
cube(x)
}
asmsm amsma 
\arguments{
  \item{x}{
kolam 
}
}
\details{
dont know
}
\value{
x numerical value
}
\references{
jisan book
}
\author{
jisan
}
\note{
no need
}


\seealso{
\mean 
}
\examples{

x<-10
sqr(x)
cube(x)
function (x) 
{
    cube <- x^3
    return(cube)
  }
}

\keyword{ square}
\keyword{ cube }
