% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try2.R
\name{MyCorr.plot}
\alias{MyCorr.plot}
\title{Visualize correlation between two numerical variable, by scatter plot}
\usage{
MyCorr.plot(x, y)
}
\arguments{
\item{x}{numerical vector of the first variable}

\item{y}{numerical vector of the second variable}
}
\description{
Visualize correlation between two numerical variable, by scatter plot
}
