% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try2.R
\name{MyCorr}
\alias{MyCorr}
\title{Evaluate correlation between two numerical variable}
\usage{
MyCorr(x, y)
}
\arguments{
\item{x}{numerical vector of the first variable}

\item{y}{numerical vector of the second variable}
}
\description{
Evaluate correlation between two numerical variable
}
