% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{My data}
\format{
a data frame with three observations and 2 variables
\describe{
\item{x}{nummern}
\item{y}{buchstaben}
}
}
\source{
simulation
}
\usage{
mydata
}
\description{
this is my first data
}
\keyword{datasets}
