\name{mypkg-package}
\alias{mypkg-package}
\alias{mypkg}
\docType{package}
\title{
\packageTitle{mypkg}
}
\description{
\packageDescription{mypkg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mypkg}
\packageIndices{mypkg}
put the value of x 
}
\author{
\packageAuthor{mypkg}

Maintainer: \packageMaintainer{mypkg}
}
\references{
jisan book
}
square, cube
\keyword{ package }
\seealso{

}
\examples{
x<-10
sqr(x)
cube(x)
}
