% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypkg2.R
\docType{package}
\name{mypkg2}
\alias{mypkg2}
\title{mypkg2: this is my fist package}
\description{
here comes the larger description
}
\section{functions}{

mysqer() is my function to calculate the square root
}

\section{datas}{

bla
}

