% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysqrt.R
\name{mysqrt}
\alias{mysqrt}
\title{Square Root}
\usage{
mysqrt(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
vector of element-wise square root
}
\description{
Calculate square root of a vector
}
\examples{
mysqrt(c(1,2,3))
}
