% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power}
\alias{power}
\title{Title}
\usage{
power(x, i)
}
\arguments{
\item{x}{value}

\item{i}{power}
}
\value{
vector of ...
}
\description{
Title
}
\examples{
power(2, 2)
}
\seealso{
\code{\link{mysqrt}}
}
