\name{sqr}
\alias{sqr}
square  cal
\title{
square calculation
}
\description{
no need....
}
\usage{
sqr(x)
}
nnASKNSKANK
\arguments{
  \item{x}{
squqre
}
}
\details{
as avove
}
\value{
x value
}
\references{
jisan book
}
\author{
jisan
}
\note{
no need
}

skksksks

\seealso{
mean
}
\examples{
x=10
sqr(x)
## The function is currently defined as
function (x) 
{
    sqr <- x^2
    return(sqr)
  }
}

\keyword{ sqruare }
\keyword{ cube }
